Add files via upload
Pizza Project where a store owner has order, stock and staff information and would like to cost of making the pizza including the ingredients cost and employee cost. Also wanted to be alerted when to order new stock and staff availability to work. Created database and normalized the data. JOINED tables for analysis using SUBQUERIES and creating VIEWS